refactor(screenshots): Move screenshot logic into main store
Screenshot logic didn't necessitate its own module.
