+++ src/shortcuts.js
@@ -1,6 +1,6 @@
 export default [
   { key: 'ctrl+o', action: 'files/promptForFiles' },
-  { key: 'ctrl+g', action: 'screenshots/takeScreenshot' },
+  { key: 'ctrl+g', action: 'takeScreenshot' },
   { key: 'r', action: 'resetActiveCamera' },
   { key: '.', action: 'increaseSlice' },
   { key: ',', action: 'decreaseSlice' },
