--- src/store/screenshots.js
@@ -1,39 +1,0 @@
-import viewHelper from 'paraview-glance/src/components/core/VtkView/helper';
-
-export default (proxyManager) => ({
-  namespaced: true,
-
-  state: {
-    showDialog: false,
-    currentScreenshot: null,
-  },
-
-  mutations: {
-    openScreenshotDialog(state, screenshot) {
-      state.currentScreenshot = screenshot;
-      state.showDialog = true;
-    },
-    closeScreenshotDialog(state) {
-      state.showDialog = false;
-    },
-  },
-
-  actions: {
-    takeScreenshot({ commit, rootState }, viewToUse = null) {
-      const view = viewToUse || proxyManager.getActiveView();
-      const viewType = viewHelper.getViewType(view);
-      if (view) {
-        return view.captureImage().then((imgSrc) => {
-          commit('openScreenshotDialog', {
-            imgSrc,
-            viewName: view.getName(),
-            viewData: {
-              background: rootState.views.backgroundColors[viewType],
-            },
-          });
-        });
-      }
-      return Promise.resolve();
-    },
-  },
-});
